+++ packages/bluesnap-direct-integration/src/index.ts
@@ -1,1 +1,2 @@
 export { default as createBlueSnapDirectCreditCardPaymentStrategy } from './create-bluesnap-direct-credit-card-payment-strategy';
+export { default as createBlueSnapDirectEcpPaymentStrategy } from './create-bluesnap-direct-ecp-payment-strategy';