+++ packages/braintree-integration/src/index.ts
@@ -1,5 +1,9 @@
 /**
  * BraintreePayPal strategies
  */
+
+export { default as createBraintreePaypalAchPaymentStrategy } from './braintree-paypal-ach/create-braintree-paypal-ach-payment-strategy';
 export { WithBraintreePaypalCustomerInitializeOptions } from './braintree-paypal/braintree-paypal-customer-options';
+
 export { default as createBraintreePaypalCustomerStrategy } from './braintree-paypal/create-braintree-paypal-customer-strategy';
+export { WithBraintreePaypalAchPaymentInitializeOptions } from './braintree-paypal-ach/braintree-paypal-ach-initialize-options';