+++ packages/core/src/payment/errors/index.ts
@@ -7,4 +7,3 @@ export {
     default as PaymentInvalidFormError,
     PaymentInvalidFormErrorDetails,
 } from './payment-invalid-form-error';
-export { default as PaymentExecuteError } from './payment-execute-error';