+++ packages/core/src/payment/payment.ts
@@ -1,4 +1,9 @@
-import { WithAccountCreation } from '@bigcommerce/checkout-sdk/payment-integration-api';
+import {
+    BlueSnapDirectEcpInstrument,
+    BlueSnapDirectEcpPayload,
+    WithAccountCreation,
+    WithBankAccountInstrument,
+} from '@bigcommerce/checkout-sdk/payment-integration-api';
 
 import { BrowserInfo } from '../common/browser-info';
 import { Omit } from '../common/types';
@@ -13,6 +18,7 @@ export default interface Payment {
 }
 
 export type PaymentInstrument =
+    | BlueSnapDirectEcpInstrument
     | CreditCardInstrument
     | (CreditCardInstrument & WithHostedFormNonce)
     | (CreditCardInstrument & WithDocumentInstrument)
@@ -23,6 +29,7 @@ export type PaymentInstrument =
     | FormattedPayload<
           | AdyenV2Instrument
           | AppleInstrument
+          | BlueSnapDirectEcpPayload
           | BoltInstrument
           | PaypalInstrument
           | FormattedHostedInstrument
@@ -41,7 +48,8 @@ export type PaymentInstrument =
     | ThreeDSVaultedInstrument
     | VaultedInstrument
     | (VaultedInstrument & WithHostedFormNonce)
-    | WithAccountCreation;
+    | WithAccountCreation
+    | WithBankAccountInstrument;
 
 export interface PaymentInstrumentMeta {
     deviceSessionId?: string;