+++ packages/core/src/payment/strategies/braintree/braintree-sdk-creator.ts
@@ -91,6 +91,7 @@ export default class BraintreeSDKCreator {
             const paypalSdkLoadCallback = () => onSuccess(braintreePaypalCheckout);
             const paypalSdkLoadConfig = {
                 currency: config.currency,
+                ...(config.isCreditEnabled && { 'enable-funding': 'paylater' }),
                 components: PAYPAL_COMPONENTS.toString(),
                 intent: config.intent,
             };