+++ packages/payment-integration-api/src/errors/index.ts
@@ -24,3 +24,5 @@ export {
 export { default as PaymentMethodInvalidError } from './payment-method-invalid-error';
 export { default as StandardError } from './standard-error';
 export { default as TimeoutError } from './timeout-error';
+export { isCustomError } from './custom-error';
+export { default as PaymentExecuteError } from './payment-execute-error';